refactor: partial refactor ipcListeners + main redux
