+++ electron/app/constants/ipcEvents.ts
@@ -1,0 +1,8 @@
+export const DOWNLOAD_PLUGIN = 'download-plugin';
+export const DOWNLOAD_PLUGIN_RESULT = 'download-plugin-result';
+export const DOWNLOAD_TEMPLATE = 'download-template';
+export const DOWNLOAD_TEMPLATE_RESULT = 'download-template-result';
+export const DOWNLOAD_TEMPLATE_PACK = 'download-template-pack';
+export const DOWNLOAD_TEMPLATE_PACK_RESULT = 'download-template-pack-result';
+export const UPDATE_EXTENSIONS = 'update-extensions';
+export const UPDATE_EXTENSIONS_RESULT = 'update-extensions-result';
