+++ electron/app/constants/telemetry.ts
@@ -1,0 +1,1 @@
+export const UPDATE_APPLICATION = 'UPDATE_APPLICATION';
