+++ electron/app/models/appconfig.ts
@@ -1,0 +1,9 @@
+export enum NewVersionCode {
+  Errored = -2,
+  NotAvailable = -1,
+  Idle = 0,
+  Checking = 1,
+  Available = 2,
+  Downloading = 3,
+  Downloaded = 4,
+}
