+++ electron/app/models/index.ts
@@ -1,0 +1,7 @@
+export interface ProjectNameChange {
+  newName: string;
+}
+
+export enum StorePropagation {
+  ChangeProjectName = 1,
+}
